Added changes to product bottom sheet and fixed other glitches in order taking
